Extract command enums to module
